requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==21.2.0
cffi==1.14.6
click==8.0.1; python_version >= '3.6'
cryptography==3.4.7
environ-config==21.2.0
flask==2.0.1
greenlet==1.1.0; python_version >= '3'
gunicorn==20.1.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.20
structlog==21.1.0
werkzeug==2.0.1; python_version >= '3.6'
